+++ packages/atomic/stencil.config.ts
@@ -1,5 +1,4 @@
 import {Config} from '@stencil/core';
-import {sass} from '@stencil/sass';
 import alias from '@rollup/plugin-alias';
 import path from 'path';
 import html from 'rollup-plugin-html';
@@ -13,7 +12,6 @@ const isDevWatch: boolean =
 export const config: Config = {
   namespace: 'atomic',
   taskQueue: 'async',
-  globalStyle: 'node_modules/bootstrap/dist/css/bootstrap-grid.css',
   outputTargets: [
     {
       type: 'dist',
@@ -38,13 +36,7 @@ export const config: Config = {
   devServer: {
     reloadStrategy: 'pageReload',
   },
-  plugins: [
-    sass({
-      includePaths: ['src/scss/'],
-      injectGlobalPaths: ['src/scss/_global.scss'],
-    }),
-    inlineSvg(),
-  ],
+  plugins: [inlineSvg()],
   rollupPlugins: {
     before: [
       isDevWatch &&