+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-state.ts
@@ -1,6 +1,14 @@
 import {CategoryFacetRequest} from './interfaces/request';
 
-export type CategoryFacetSetState = Record<string, CategoryFacetRequest>;
+export type CategoryFacetSlice = {
+  request: CategoryFacetRequest;
+  initialNumberOfValues: number;
+};
+
+export type CategoryFacetSetState = Record<
+  string,
+  CategoryFacetSlice | undefined
+>;
 
 export function getCategoryFacetSetInitialState(): CategoryFacetSetState {
   return {};