+++ packages/headless/src/features/facets/facet-set/facet-set-slice.test.ts
@@ -1,5 +1,4 @@
 import {facetSetReducer, convertFacetValueToRequest} from './facet-set-slice';
-import {getHistoryEmptyState} from '../../history/history-slice';
 import {change} from '../../history/history-actions';
 import {
   registerFacet,
@@ -8,6 +7,7 @@ import {
   updateFacetSortCriterion,
   updateFacetNumberOfValues,
   updateFacetIsFieldExpanded,
+  updateFreezeCurrentValues,
 } from './facet-set-actions';
 import {buildMockFacetValue} from '../../../test/mock-facet-value';
 import {buildMockSearch} from '../../../test/mock-search';
@@ -23,6 +23,7 @@ import {selectFacetSearchResult} from '../facet-search-set/specific/specific-fac
 import * as FacetReducers from '../generic/facet-reducer-helpers';
 import {FacetSetState, getFacetSetInitialState} from './facet-set-state';
 import {deselectAllFacets} from '../generic/facet-actions';
+import {getHistoryInitialState} from '../../history/history-state';
 import {restoreSearchParameters} from '../../search-parameters/search-parameter-actions';
 
 describe('facet-set slice', () => {
@@ -99,7 +100,7 @@ describe('facet-set slice', () => {
       foo: buildMockFacetRequest(),
     };
     const historyChange = {
-      ...getHistoryEmptyState(),
+      ...getHistoryInitialState(),
       facetSet: expectedFacetSet,
     };
 
@@ -115,7 +116,7 @@ describe('facet-set slice', () => {
     const state = {foo: buildMockFacetRequest()};
     const emptyFacetSet = {};
     const historyChange = {
-      ...getHistoryEmptyState(),
+      ...getHistoryInitialState(),
       facetSet: emptyFacetSet,
     };
 
@@ -208,6 +209,29 @@ describe('facet-set slice', () => {
     const action = toggleSelectFacetValue({
       facetId: '1',
       selection: facetValue,
+    });
+
+    expect(() => facetSetReducer(state, action)).not.toThrow();
+  });
+
+  it(`dispatching #updateFreezeCurrentValues with a registered facet id
+  sets the state updateFreezeCurrentValues to the value`, () => {
+    const id = '1';
+
+    state[id] = buildMockFacetRequest();
+
+    const action = updateFreezeCurrentValues({
+      facetId: id,
+      freezeCurrentValues: true,
+    });
+    const finalState = facetSetReducer(state, action);
+    expect(finalState[id].freezeCurrentValues).toBe(true);
+  });
+
+  it('dispatching #updateFreezeCurrentValues with an invalid id does not throw', () => {
+    const action = updateFreezeCurrentValues({
+      facetId: '1',
+      freezeCurrentValues: true,
     });
 
     expect(() => facetSetReducer(state, action)).not.toThrow();