+++ packages/headless/src/features/facets/facet-set/facet-set-slice.ts
@@ -7,6 +7,7 @@ import {
   updateFacetSortCriterion,
   updateFacetNumberOfValues,
   updateFacetIsFieldExpanded,
+  updateFreezeCurrentValues,
 } from './facet-set-actions';
 import {executeSearch} from '../../search/search-actions';
 import {selectFacetSearchResult} from '../facet-search-set/specific/specific-facet-search-actions';
@@ -83,19 +84,35 @@ export const facetSetReducer = createReducer(
         facetRequest.freezeCurrentValues = true;
         facetRequest.preventAutoSelect = true;
       })
+      .addCase(updateFreezeCurrentValues, (state, action) => {
+        const {facetId, freezeCurrentValues} = action.payload;
+        const facetRequest = state[facetId];
+
+        if (!facetRequest) {
+          return;
+        }
+
+        facetRequest.freezeCurrentValues = freezeCurrentValues;
+      })
       .addCase(deselectAllFacetValues, (state, action) => {
-        handleFacetDeselectAll<FacetRequest>(state, action.payload);
+        const request = state[action.payload];
+        handleFacetDeselectAll<FacetRequest>(request);
       })
       .addCase(deselectAllFacets, (state) => {
         Object.keys(state).forEach((facetId) => {
-          handleFacetDeselectAll<FacetRequest>(state, facetId);
+          const request = state[facetId];
+          handleFacetDeselectAll<FacetRequest>(request);
         });
       })
       .addCase(updateFacetSortCriterion, (state, action) => {
         handleFacetSortCriterionUpdate<FacetRequest>(state, action.payload);
       })
       .addCase(updateFacetNumberOfValues, (state, action) => {
-        handleFacetUpdateNumberOfValues<FacetRequest>(state, action.payload);
+        const {facetId, numberOfValues} = action.payload;
+        handleFacetUpdateNumberOfValues<FacetRequest>(
+          state[facetId],
+          numberOfValues
+        );
       })
       .addCase(updateFacetIsFieldExpanded, (state, action) => {
         const {facetId, isFieldExpanded} = action.payload;