+++ packages/headless/src/features/facets/range-facets/date-facet-set/date-facet-analytics-actions.ts
@@ -1,0 +1,6 @@
+import {logRangeFacetBreadcrumb} from '../generic/range-facet-analytics-actions';
+
+/**
+ * Logs a date facet breadcrumb event.
+ */
+export const logDateFacetBreadcrumb = logRangeFacetBreadcrumb;